requirements.txt
numpy==1.23.1
networkx==3.1
torch==1.11.0+cu113
torchvision==0.12.0+cu113
mmcv-full==1.6.0
markdown-it-py==2.2.0
timm==0.4.12
opencv-python==4.7.0.72
utm==0.7.0
setuptools==69.1.0
pyproj==3.3.0
pandas==1.4.3
rasterio==1.3a3
pyhdf==0.10.2
glob2==0.7
matplotlib==3.5.2
scikit-image == 0.19.0
scikit-learn == 1.0.1
scipy==1.8.1
pyresample==1.24.1
h5py==3.7.0
tqdm==4.66.2
tensorboard==2.14.0